requirements.txt
boto3==1.18.8
python-dotenv==0.19.0
awscli==1.20.8
aws==0.2.5
pandas==1.3.1
